cleanup and creation of weakspot

+++ src/components/characterComponents/HitInformation.cs
@@ -3,15 +3,15 @@
 namespace FirstPerson.scenes.enemies.test;
 
 public class HitInformation(
-    int? healthDamage = null, 
-    int? staggerDamage = null, 
+    float? healthDamage = null, 
+    float? staggerDamage = null, 
     Vector3? sourceGlobalPosition = null,
     Vector3? collisionGlobalPosition = null,
     float? pitch = null,
     float? roll = null)
 {
-    public int? healthDamage = healthDamage;
-    public int? staggerDamage = staggerDamage;
+    public float? healthDamage = healthDamage;
+    public float? staggerDamage = staggerDamage;
     public Vector3? sourceGlobalPosition = sourceGlobalPosition;
     public Vector3? collisionGlobalPosition = collisionGlobalPosition;
     public float? pitch = pitch;
